shared/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL
mysqlclient==2.2.0  # MySQL
aiosqlite==0.19.0  # SQLite

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.5

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Queue Management
redis==5.0.1
celery==5.3.4

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
boto3==1.29.3  # AWS SDK

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
tenacity==8.2.3  # Retry logic 